requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0  # Added for Render deployment
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic[email]==2.4.2  # Includes email-validator
python-dotenv==1.0.0
alembic==1.12.1
jinja2==3.1.2
python-multipart==0.0.6
email-validator==2.1.0.post1  # Required for EmailStr validation

# AI and LLM
openai==1.68.2  # Updated version
crewai==0.108.0 # Updated version

# Quality checks
language-tool-python==2.7.1
spacy==3.8.4 # Updated version

# Integrations
stripe==7.0.0
requests==2.31.0
tenacity==8.2.3
sendgrid==6.11.0 # Added for email delivery

# Testing
pytest==7.4.3
httpx==0.27.2 # Updated version
